fix: 기존 private에서 공개 공유페이지로 변경
