+++ src/app/(private)/study-rooms/[id]/homework/[homeworkId]/page.tsx
@@ -1,45 +1,8 @@
-import type { Metadata } from 'next';
-import { headers } from 'next/headers';
-
-import { SITE_CONFIG } from '@/config/site';
 import HomeworkDetailRoleSwitch from '@/features/homework/components/detail/homework-detail-role-switch';
 
 type Props = {
   params: Promise<{ id: string; homeworkId: string }>;
 };
-
-export async function generateMetadata({ params }: Props): Promise<Metadata> {
-  const { id, homeworkId } = await params;
-  const origin = await getRequestOrigin();
-  const url = `${origin}/study-rooms/${id}/homework/${homeworkId}`;
-  const imageUrl = `${url}/opengraph-image`;
-  const title = `과제 | ${SITE_CONFIG.name}`;
-
-  return {
-    title,
-    alternates: { canonical: url },
-    openGraph: {
-      locale: 'ko_KR',
-      type: 'article',
-      title,
-      url,
-      siteName: SITE_CONFIG.name,
-      images: [
-        {
-          url: imageUrl,
-          width: 1200,
-          height: 630,
-          alt: '과제',
-        },
-      ],
-    },
-    twitter: {
-      card: 'summary_large_image',
-      title,
-      images: [imageUrl],
-    },
-  };
-}
 
 export default async function HomeworkDetailPage({ params }: Props) {
   const resolvedParams = await params;
@@ -53,11 +16,3 @@ export default async function HomeworkDetailPage({ params }: Props) {
     />
   );
 }
-
-const getRequestOrigin = async () => {
-  const headersList = await headers();
-  const host = headersList.get('x-forwarded-host') ?? headersList.get('host');
-  const protocol = headersList.get('x-forwarded-proto') ?? 'https';
-
-  return host ? `${protocol}://${host}` : SITE_CONFIG.url;
-};
