+++ src/app/(private)/study-rooms/[id]/note/[noteId]/page.tsx
@@ -1,7 +1,3 @@
-import type { Metadata } from 'next';
-import { headers } from 'next/headers';
-
-import { SITE_CONFIG } from '@/config/site';
 import BackLink from '@/features/dashboard/studynote/components/back-link';
 import { StudyNoteDetailContentsSection } from '@/features/dashboard/studynote/detail/components/contents-section';
 import { StudyNoteDetailMetaSection } from '@/features/dashboard/studynote/detail/components/meta-section';
@@ -12,39 +8,6 @@ import { PageViewTracker } from '@/shared/components/gtm';
 type Props = {
   params: Promise<{ id: string; noteId: string }>;
 };
-
-export async function generateMetadata({ params }: Props): Promise<Metadata> {
-  const { id, noteId } = await params;
-  const origin = await getRequestOrigin();
-  const url = `${origin}/study-rooms/${id}/note/${noteId}`;
-  const imageUrl = `${url}/opengraph-image`;
-  const title = `수업노트 | ${SITE_CONFIG.name}`;
-
-  return {
-    title,
-    alternates: { canonical: url },
-    openGraph: {
-      locale: 'ko_KR',
-      type: 'article',
-      title,
-      url,
-      siteName: SITE_CONFIG.name,
-      images: [
-        {
-          url: imageUrl,
-          width: 1200,
-          height: 630,
-          alt: '수업노트',
-        },
-      ],
-    },
-    twitter: {
-      card: 'summary_large_image',
-      title,
-      images: [imageUrl],
-    },
-  };
-}
 
 export default async function StudyNoteDetailPage({ params }: Props) {
   const resolvedParams = await params;
@@ -68,11 +31,3 @@ export default async function StudyNoteDetailPage({ params }: Props) {
     </>
   );
 }
-
-const getRequestOrigin = async () => {
-  const headersList = await headers();
-  const host = headersList.get('x-forwarded-host') ?? headersList.get('host');
-  const protocol = headersList.get('x-forwarded-proto') ?? 'https';
-
-  return host ? `${protocol}://${host}` : SITE_CONFIG.url;
-};
